Separate logic of ProfilerWhatChanged to a component
