+++ packages/react-devtools-shared/src/devtools/views/Components/ProfilerWhatChanged.css
@@ -1,0 +1,30 @@
+.Component {
+  margin-bottom: 1rem;
+}
+
+.Item {
+  margin-top: 0.25rem;
+}
+
+.Key {
+  font-family: var(--font-family-monospace);
+  font-size: var(--font-size-monospace-small);
+  line-height: 1;
+}
+
+.Key:first-of-type::before {
+  content: ' (';
+}
+
+.Key::after {
+  content: ', ';
+}
+
+.Key:last-of-type::after {
+  content: ')';
+}
+
+.Label {
+  font-weight: bold;
+  margin-bottom: 0.5rem;
+}
